Do not delete node_modules if it is part of test assets
Previously node_modules was always deleted, making successive runs of the
"NPM aborts when the node_modules directory already exists" test fail. Fix
this by not deleting node_modules at all if it is present from the start,
which is good behavior anyway.

Additionally, make the test case interceptor only delete node_modules if
there is no .gitkeep inside.

+++ provenance-analyzer/src/funTest/kotlin/NpmTest.kt
@@ -30,7 +30,8 @@ class NpmTest : StringSpec() {
             projectBaseDir.listFiles().forEach {
                 if (it.isDirectory) {
                     val nodeModulesDir = File(it, "node_modules")
-                    if (nodeModulesDir.isDirectory) {
+                    val gitKeepFile = File(nodeModulesDir, ".gitkeep")
+                    if (nodeModulesDir.isDirectory && !gitKeepFile.isFile) {
                         nodeModulesDir.deleteRecursively()
                     }
                 }
